pyproject.toml
[project]
name = "python-or-kr-wip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "celery>=5.5.2",
    "django>=5.1.7",
    "django-celery-beat>=2.8.1",
    "django-celery-results>=2.6.0",
    "django-tailwind[reload]>=3.8.0",
    "feedparser>=6.0.11",
    "google-generativeai>=0.8.5",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "langdetect>=1.0.9",
    "llm>=0.24.2",
    "llm-gemini>=0.18.1",
    "logfire[celery,django,psycopg,system-metrics]>=3.16.0",
    "lxml>=5.3.2",
    "markdown>=3.7",
    "markitdown[all]>=0.1.2",
    "psycopg[binary]>=3.2.5",
    "pydantic-ai-slim[anthropic,openai]>=0.2.6",
    "pydantic-ai[logfire]>=0.2.6",
    "pytz>=2025.2",
    "readtime>=3.0.0",
    "requests>=2.32.3",
    "tiktoken>=0.9.0",
    "trafilatura>=2.0.0",
    "wagtail>=6.4.1",
    "wagtail-bakery>=0.8.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "pythonkr_backend.settings"
pythonpath = ["pythonkr_backend"]
testpaths = [
    "pythonkr_backend/curation",
    "pythonkr_backend/pythonkr",
]
python_files = ["tests.py", "test_*.py", "*_tests.py"]

[dependency-groups]
dev = [
    "pytest-django>=4.10.0",
    "ruff>=0.9.10",
]
